Metrics: correct wal_receiver.replay_lag_b datatype to int8
Enchances PR #335

+++ pgwatch2/sql/config_store/metric_definitions.sql
@@ -5268,7 +5268,7 @@ true,
 $sql$
 select
   (extract(epoch from now()) * 1e9)::int8 as epoch_ns,
-  pg_xlog_location_diff(pg_last_xlog_receive_location(), pg_last_xlog_replay_location()) as replay_lag_b,
+  pg_xlog_location_diff(pg_last_xlog_receive_location(), pg_last_xlog_replay_location())::int8 as replay_lag_b,
   extract(epoch from (now() - pg_last_xact_replay_timestamp()))::int8 as last_replay_s;
 $sql$,
 '{"prometheus_all_gauge_columns": true}'
@@ -5282,7 +5282,7 @@ true,
 $sql$
 select
   (extract(epoch from now()) * 1e9)::int8 as epoch_ns,
-  pg_wal_lsn_diff(pg_last_wal_receive_lsn(), pg_last_wal_replay_lsn()) as replay_lag_b,
+  pg_wal_lsn_diff(pg_last_wal_receive_lsn(), pg_last_wal_replay_lsn())::int8 as replay_lag_b,
   extract(epoch from (now() - pg_last_xact_replay_timestamp()))::int8 as last_replay_s;
 $sql$,
 '{"prometheus_all_gauge_columns": true}'
